src/app/requirements.txt
langchain==0.3.13
fastapi==0.115.6
autogen-agentchat==0.2.40
spotipy==2.25.0
uvicorn==0.34.0
streamlit==1.42.0
groq==0.18.0